+++ recetas/delete.php
@@ -5,7 +5,7 @@ verify_csrf();
 
 $id = (int) ($_POST['id'] ?? 0);
 if ($id) {
-    db()->prepare('DELETE FROM recetas WHERE id = ?')->execute([$id]);
+    db()->prepare('DELETE FROM recetas WHERE id = ? AND consultorio_id = ?')->execute([$id, tenant_id()]);
     flash('Receta eliminada.');
 }
 redirect('/recetas/index.php');
